move the dictionary search stuff into `src/dictionary/dictionarySearchContainer.tsx`
